+++ src/DateTimeFactory.php
@@ -5,6 +5,7 @@ declare(strict_types = 1);
 namespace Avtocod\B2BApi;
 
 use DateTime;
+use DateTimeInterface;
 use InvalidArgumentException;
 
 /**
@@ -37,11 +38,11 @@ class DateTimeFactory extends DateTime
     /**
      * Convert DateTime object into string, using ISO8601 (zulu) format.
      *
-     * @param DateTime $date_time
+     * @param DateTimeInterface $date_time
      *
      * @return string E.g.: '2017-01-05T16:45:23.000Z'
      */
-    public static function toIso8601Zulu(DateTime $date_time): string
+    public static function toIso8601Zulu(DateTimeInterface $date_time): string
     {
         return $date_time->format('Y-m-d\\TH:i:s.v\\Z');
     }