+++ src/Responses/UserReportTypesResponse.php
@@ -4,9 +4,9 @@ declare(strict_types = 1);
 
 namespace Avtocod\B2BApi\Responses;
 
-use DateTime;
 use Countable;
 use ArrayIterator;
+use DateTimeImmutable;
 use IteratorAggregate;
 use Tarampampam\Wrappers\Json;
 use Avtocod\B2BApi\DateTimeFactory;
@@ -36,7 +36,7 @@ class UserReportTypesResponse implements ResponseInterface, Countable, IteratorA
     protected $size;
 
     /**
-     * @var DateTime
+     * @var DateTimeImmutable
      */
     protected $stamp;
 
@@ -56,14 +56,14 @@ class UserReportTypesResponse implements ResponseInterface, Countable, IteratorA
      * @param string                 $raw_response
      * @param string                 $state
      * @param int                    $size
-     * @param DateTime               $stamp
+     * @param DateTimeImmutable      $stamp
      * @param array<int, ReportType> $data
      * @param int|null               $total
      */
     private function __construct(string $raw_response,
                                  string $state,
                                  int $size,
-                                 DateTime $stamp,
+                                 DateTimeImmutable $stamp,
                                  array $data,
                                  ?int $total)
     {
@@ -104,7 +104,7 @@ class UserReportTypesResponse implements ResponseInterface, Countable, IteratorA
             $raw_response,
             $as_array['state'],
             $as_array['size'],
-            DateTimeFactory::createFromIso8601Zulu($as_array['stamp']),
+            DateTimeImmutable::createFromMutable(DateTimeFactory::createFromIso8601Zulu($as_array['stamp'])),
             $as_array['data'],
             $as_array['total'] ?? null
         );
@@ -133,9 +133,9 @@ class UserReportTypesResponse implements ResponseInterface, Countable, IteratorA
     /**
      * Get response date/time.
      *
-     * @return DateTime
+     * @return DateTimeImmutable
      */
-    public function getStamp(): DateTime
+    public function getStamp(): DateTimeImmutable
     {
         return $this->stamp;
     }