+++ src/Responses/UserResponse.php
@@ -4,9 +4,9 @@ declare(strict_types = 1);
 
 namespace Avtocod\B2BApi\Responses;
 
-use DateTime;
 use Countable;
 use ArrayIterator;
+use DateTimeImmutable;
 use IteratorAggregate;
 use Tarampampam\Wrappers\Json;
 use Avtocod\B2BApi\DateTimeFactory;
@@ -36,7 +36,7 @@ class UserResponse implements WithRawResponseGetterInterface, ResponseInterface,
     protected $size;
 
     /**
-     * @var DateTime
+     * @var DateTimeImmutable
      */
     protected $stamp;
 
@@ -48,13 +48,13 @@ class UserResponse implements WithRawResponseGetterInterface, ResponseInterface,
     /**
      * Create a new response instance.
      *
-     * @param string           $raw_response
-     * @param string           $state
-     * @param int              $size
-     * @param DateTime         $stamp
-     * @param array<int, User> $data
+     * @param string            $raw_response
+     * @param string            $state
+     * @param int               $size
+     * @param DateTimeImmutable $stamp
+     * @param array<int, User>  $data
      */
-    private function __construct(string $raw_response, string $state, int $size, DateTime $stamp, array $data)
+    private function __construct(string $raw_response, string $state, int $size, DateTimeImmutable $stamp, array $data)
     {
         $this->raw_response_content = $raw_response;
         $this->state                = $state;
@@ -92,7 +92,7 @@ class UserResponse implements WithRawResponseGetterInterface, ResponseInterface,
             $raw_response,
             $as_array['state'],
             $as_array['size'],
-            DateTimeFactory::createFromIso8601Zulu($as_array['stamp']),
+            DateTimeImmutable::createFromMutable(DateTimeFactory::createFromIso8601Zulu($as_array['stamp'])),
             $as_array['data']
         );
     }
@@ -120,9 +120,9 @@ class UserResponse implements WithRawResponseGetterInterface, ResponseInterface,
     /**
      * Get response date/time.
      *
-     * @return DateTime
+     * @return DateTimeImmutable
      */
-    public function getStamp(): DateTime
+    public function getStamp(): DateTimeImmutable
     {
         return $this->stamp;
     }