+++ tests/Unit/Responses/Entities/ReportMadeTest.php
@@ -4,6 +4,7 @@ declare(strict_types = 1);
 
 namespace Avtocod\B2BApi\Tests\Unit\Responses\Entities;
 
+use DateTimeImmutable;
 use Avtocod\B2BApi\DateTimeFactory;
 use Avtocod\B2BApi\Tests\AbstractTestCase;
 use Avtocod\B2BApi\Responses\Entities\ReportMade;
@@ -25,7 +26,7 @@ class ReportMadeTest extends AbstractTestCase
             $report_uid = $attributes['report_uid'],
             $is_new = $attributes['is_new'],
             $process_request_uid = $attributes['process_request_uid'],
-            $suggest_get = DateTimeFactory::createFromIso8601Zulu($attributes['suggest_get'])
+            $suggest_get = DateTimeImmutable::createFromMutable(DateTimeFactory::createFromIso8601Zulu($attributes['suggest_get']))
         );
 
         $this->assertSame($report_uid, $made->getReportUid());